refactor(integrations): some import changes
