+++ crates/services/integration/src/emby.rs
@@ -1,16 +1,15 @@
+use enums::{MediaLot, MediaSource};
+use media_models::{IntegrationMediaCollection, IntegrationMediaSeen};
+use rust_decimal::Decimal;
 use rust_decimal_macros::dec;
 use sea_orm::prelude::async_trait::async_trait;
 use sea_orm::DatabaseConnection;
+use serde::{Deserialize, Serialize};
 
-use enums::{MediaLot, MediaSource};
-use media_models::{IntegrationMediaCollection, IntegrationMediaSeen};
-
-use crate::integration::Integration;
-use crate::show_identifier::ShowIdentifier;
+use super::{integration::YankIntegration, show_identifier::ShowIdentifier};
 
 mod models {
-    use rust_decimal::Decimal;
-    use serde::{Deserialize, Serialize};
+    use super::*;
 
     #[derive(Serialize, Deserialize, Debug, Clone)]
     #[serde(rename_all = "PascalCase")]
@@ -117,8 +116,8 @@ impl ShowIdentifier for EmbyIntegration {
     }
 }
 
-impl Integration for EmbyIntegration {
-    async fn progress(
+impl YankIntegration for EmbyIntegration {
+    async fn yank_progress(
         &self,
     ) -> anyhow::Result<(Vec<IntegrationMediaSeen>, Vec<IntegrationMediaCollection>)> {
         self.emby_progress().await
